[BUGFIX] Fix JavaScript error for read-only content elements
Currently a JavaScript error is thrown when accessing a page with
the Page module containing a content element which the backend
user doesn't have editing right for. This happens because the drop
zones uses the title of the content element icon as the title for
the drop zone. However when a element is not editable there is no
link wrapped around the icon and thus the selector doesn't find
the icon.

Change-Id: If0943b040bbaf761282e4e4757ad4e565e910e44
Fixes: #48675
Reviewed-on: https://review.typo3.org/21168
Reviewed-by: Aske Ertmann
Tested-by: Aske Ertmann
Reviewed-by: Jo Hasenau
Tested-by: Jo Hasenau