[BUGFIX] Show only translated subset of children for language "all"
This patch fixes the behaviour of the columns view. It will now
show only those children of a container that are matching the
currently selected language of the columns view.
So containers with language "all" will not show all their children
anymore.

Additionally it fixes the problem that translations of copied records
lost their sys_language_uid, so they didn't show up in the correct
parent containers.

And the paste icons have been fixed so copies of translated records
won't run into errors anymore.

Change-Id: I02218c1c19374429f260dd4ff0d3ca74228f14f5
Resolves: #53549
Branches: 3.0, 2.1
Reviewed-on: https://review.typo3.org/25900
Reviewed-by: Fabien Udriot
Reviewed-by: Eric Chavaillaz
Tested-by: Eric Chavaillaz
Reviewed-by: Jo Hasenau
Tested-by: Jo Hasenau