[BUGFIX] Show only translated subset of children for language "all"
This patch fixes the behaviour of the columns view. It will now
show only those children of a container that are matching the
currently selected language of the columns view.
So containers with language "all" will not show all their children
anymore.

Additionally it fixes the problem that translations of copied records
lost their sys_language_uid, so they didn't show up in the correct
parent containers.

And the paste icons have been fixed so copies of translated records
won't run into errors anymore.

Change-Id: Iecb770d41334b523c9dd4579e086becda273a38e
Resolves: #53549
Branches: 2.1