[BUGFIX] Add necessary CSS classes to page columns in language view
Due to the fact that the language view is rendered completely different
from the default page view, we have to adapt the Gridelements JS files
to provide the information that the core does not offer.

To get the allowed CE feature, we need to know about the position of each
column. The patch adds the colPos value and creates the t3-allow classes
to language columns as well now.

Change-Id: If620b1fc330f3b70671027e7abf9562eea3c231f
Resolves: #55620
Reviewed-on: https://review.typo3.org/27311
Reviewed-by: Jo Hasenau
Tested-by: Jo Hasenau