[BUGFIX] Load pi_flexform data structure inside of the backend utility
Loading the tt_content::pi_flexform data structure inside of the
FormEngine for rendering and on PreProcessFieldArray for writing
isn't a good idea.

The FormEngine step is too early to have the configuration in
place when existing form data gets loaded, so relation traversing
cannot be done.

Loading a single data structure in the GLOBALS[TCA] just to make
it available for writing isn't a good idea, too. In case someone
creates something like nested tt_contents of different types
and pushes a bunch of them to bulk-update to TCEmain, overwriting
the global field might cause conflicts.

Change-Id: I399ee569ed9d2e83e3f71701f99285a5725ddebe
Relates: #56379
Reviewed-on: https://review.typo3.org/27906
Reviewed-by: Jo Hasenau
Tested-by: Jo Hasenau