[BUGFIX] Reenable moving to first column positions after regression fix
It introduces the method that caused the regression again, but makes
sure that it will not affect any children as before.

Translated records are treated separately before the hook,
thus we must make sure to update them as well, since there is no element
to move them "after" so they can't get colPos and sorting values.

Change-Id: I4c20f0dfd124d330d98082a8c0b3760666720222
Resolves: #71801
Releases: 3-0
Reviewed-on: https://review.typo3.org/44928
Reviewed-by: Jo Hasenau <info@cybercraft.de>
Tested-by: Jo Hasenau <info@cybercraft.de>