[BUGFIX] Make layout setup aware of given icon file paths again
Rules are:

You can provide iconIdentifier and iconIdentifierLarge yourself,
but then you have to register the icon beforehand and any icon
file you provided in the setup will be ignored.

If you provide an icon file in the setup, the identifiers are
generated automatically, so you must not set them manually.

Change-Id: I6f90a0be5b8fc980ac4cf5ed932045f7e081aa5f
Resolves: #73198
Releases: master, 7-0
Reviewed-on: https://review.typo3.org/46644
Reviewed-by: Jo Hasenau <info@cybercraft.de>
Tested-by: Jo Hasenau <info@cybercraft.de>