[BUGFIX] Fixes checking reference index when gridelements 7.0.6-dev is installed.
The error was caused by a php type hint added to the first parameter of
the hook method as the given value is not always an array. 
For example if the calling BackendUtility class gets an error 
when parsing the flexform xml, then the given value is only a string.

Resolves: #80200

Change-Id: If65a7520708caab08e18f49d8acd94787fe7c901
Reviewed-on: https://review.typo3.org/51982
Reviewed-by: Jo Hasenau <info@cybercraft.de>
Tested-by: Jo Hasenau <info@cybercraft.de>