[BUGFIX] Fix wrong name for context menu javascript
The wrong naming here breaks the functionality 'paste as reference after' within the context menu when copying a content element and clicking on the icon of another content element for pasting a reference of the copy after the element. The file Resources/Public/JavaScript/ClickMenuActions.js does not exist, I think this is just a logical error.